Require that only identifiers and member expressions may be extended
Summary:
This diff syntactically locks down the set of expressions that may appear in superclass position to variables, and member expressions that ultimately are rooted in a variable and that have only property identifier lookups.

This is required because the super of a class, which is syntactically at the moment an arbitrary expression, is typechecked as part of the class's signature, which we assume is a type-like structure and can be typechecked without resolving all tvars. However, if the super is an arbitrary expression, we might perform operations that we want to require to have a resolved tvar for such as function calls on them, before the variable is actually resolved.

The errors here are most apparent when we seal tvars at expression boundaries, because then cases like

Reviewed By: SamChou19815

Differential Revision: D39913510

fbshipit-source-id: 9e95b83e8cde6dcf1fc62e6eaf48aff295c3aa3f