Cycle errors and codemod handling of synthesizable object literals
Summary:
Previously we made object literals whose fields are all annotated functions count as being annotated themselves for the purpose of cycle detection. However, when reporting cycle errors or codemodding, if an object is in a cycle, and only contains functions, but some of those functions are unannotated, we'll still only ask for the object's variable as a whole to be annotated.

With this diff, we now point to the missing return type annotations for such objects, and modify the codemod accordingly. This will help avoid adding/requiring large complicated annotations on declarations.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40153994

fbshipit-source-id: 7a92c864ff0ca72e7badf0de4bfaeb14839d1696