Use a fresh constraint cache where we might reset errors
Summary:
When flow encounters the same constraint already cached, it won't do anything. In the past, this is fine, since running constraints are idempotent.

However, in a few places we will now reset errors. This breaks the idempotent assumption. Sometimes we reset errors because we know in the checking time, we will eventually produce the same error. If we don't use a fresh constraint cache, in some cases, the synthesis flow and the checking flow are the same and they both introduce errors, we will miss the error in the end.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D40128757

fbshipit-source-id: f9256f742527a1c0b796877a4bea9a85c2fbac5a