Add run_in_post_inference_mode function to context
Summary:
I'd like to use the tvar_resolver to resolve inputs to the implicit instantiation post-pass. But I'd still like to be able to distinguish between being in the checking vs. post-inference phase. `phase` on context was immutable, so I made it mutable and added a function to run in that mode.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40191471

fbshipit-source-id: 98550c6d023c00421a1f47e874345d5271570d02