Improved support for annotated this
Summary:
This diff implements support for this-annotations within the name resolver and env builder. This addresses an issue where functions with annotated `this` types are considered to be self-cyclical and require annotations on their return type under LTI.

Annotated thises are now modeled as standard writes, and in contexts where we need to do work for unannotated thises, we now ensure that the this annotation isn't present.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40165128

fbshipit-source-id: a8063e0b46bfdeb1267b7c5fb2d80c8e798215f6