[reordered-env] Non-assigning type writes should never be added as dependents
Summary:
Previously, we will unconditionally add dep for type identifiers, because it will always be in the environment. This is a bad assumption to begin with, since for the following program:

```
type A = ...
type A = ...
```

only one of them will be there. However, it doesn't run into problems before, because we will create name-def for both legal and illegal writes.

This diff fixes the problem by never adding non-assigning type writes as dependents.

Reviewed By: mvitousek

Differential Revision: D40286793

fbshipit-source-id: 6ccb77e0985662cff3b6b4e717a74bcdb5775fd2