Create tarjan nodes on demand
Summary:
Tarjan's strongly-connected component algorithm maintains some state associated
with each vertex of the graph. Before this diff, we would initialize the
traversal by lifting the graph of values into a graph of nodes, where the nodes
include this state.

The algorithm performs a traversal starting at a provided root set. During
recheck, the root set contains the direct dependencies of changed files and
their transitive dependents. For many rechecks, the traversal starting with
these roots will only visit a small portion of the graph.

Before this change, we would create nodes for the entire graph up front. Per the
above, we would not necessarily visit those nodes. This change makes it cheaper
to initialize the traversal by creating nodes on demand instead.

For very large repos, creating O(repo) nodes gets pretty expensive. Most
rechecks have small fan-out, so we can avoid most of this work most of the time
in practice.

I think we could be more clever here and make this even faster, but this is a
nice improvement for not much effort.

Reviewed By: mroch

Differential Revision: D40286671

fbshipit-source-id: 4f924b1e0c88181763c3a6d502bd1a324ab43c8b