fix clock when reconnecting to watchman
Summary:
when we lose our connection to watchman, we try to reconnect and ask it if it's a fresh instance. if it's not a fresh instance, then we can resubscribe and use the previous clock to find out what we missed while we were disconnected.

if it is a fresh instance, then watchman itself missed changes, so we go into a recovery mode where we resubscribe and then ask for all of the changes since the mergebase. however, we were still subscribing with the previous clock, so the new subscription would eventually return is_fresh_instance = true, and we'd think the _new_ subscription missed updates.

the fix is to not resubscribe with the previous clock when we're handling missed changes. since we're already handling the missed changes, we don't need nor want the changes since the previous clock.

Changelog: [fix] Fix a bug that can cause the Flow server to unnecessarily restart when Watchman restarts

Reviewed By: samwgoldman

Differential Revision: D40316477

fbshipit-source-id: 097f962893febfc481932bf91d4165f5def4ee86