fix mergebase tracking when reconnecting to watchman
Summary:
when we lose our connection to watchman, we re-init, which creates a new env. the env keeps track of the mergebase, and we even query for the mergebase after the re-init, but weren't saving that mergebase in the env!

without this info, we can't decide whether to restart to save time, and can't decide if we can recover from missed changes.

Changelog: [fix] Fix a bug that can cause the Flow server to unnecessarily restart when Watchman restarts

Reviewed By: samwgoldman

Differential Revision: D40315674

fbshipit-source-id: 584a8c359b416c5196ea94dbb3bb44e8e318514e