Fix destructuring parameters that bind no variables
Summary:
Currently in LTI, when a function parameter binds no variables due to destructuring--e.g. `function({}) { }`--we report an error with a read of an unprepared tvar pointing to the destructured parameter. This is the read of the FunctionParamLoc, which is normally populated when we check the component that contains any function parameters, destructured or otherwise. Since there is no such component in this case, we never prepare this location for reading. In addition, due to this fact, any attempt to apply such a function will treat the function parameter as `empty`, rather than e.g. an empty object type.

This diff fixes this by introducing a new name_def used only when a parameter binds no names. We determine that the parameter binds no names by making the name_def destructuring system work as a fold (while maintaining a unitary interface to it for use elsewhere), so that we know if any names were ever bound during a destructuring. This name_def is resolved to any, which matches [Flow's current behavior](https://flow.org/try/#0BTDaF0EoAIF4D5 (https://github.com/facebook/flow/commit/511c6024d110bffb7ede2472656289c1d76f654b)oBYBMlgEYAMXJA) (though there may be some coverage differences due to any propagation, since previously the unsoundness here was due to open tvars, not the any type). This name_def will make it relatively straightforward to improve the soundness here in the future (by including the pattern in the name def and converting it to the appropriate object or array type).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40321637

fbshipit-source-id: 48d4162b7bf552f3bd68336c74887bc5450f1f70