Fix ordering for classes
Summary:
We currently override the class visit method in name_def_ordering, which causes all classes to only depend on their signatures. We can do so because we know they'll be resolved during env_resolution, and will eventually be checked against their signatures. However, this isn't the case for class expressions since they're checked by `Statement.expression`. This diff simply removes the override: we still determine ordering for classes that are in their own def by examining only the signature, but for other classes we'll depend on their entire bodies.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40322632

fbshipit-source-id: b88ab1317e40161e72cc7891421090b2f1c8a702