[contextual] Use singleton speculation to power hint eval error rollback
Summary:
In this diff, I used speculation mechanism to implement the logic in hint decomposition to detect whether we have an error and then rollback potential errors.

I used the existing intersection cases spec, but special case the behavior when there is only one element (impossible case in normal intersection types). When there is only one case in intersection case and it fails for whatever reasons, it will throw an exception. type_hint will use this exception to learn that some errors will be produced in decomposition and abort.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D40170581

fbshipit-source-id: 6b7515912f981ae70bc62cd2c1d9ca899e5fcc7e