[contextual] First cut of multi-spec for hint decomposition
Summary:
In this diff, I use the new singleton speculation machinery to implement the logic of keeping successful flows in union type, which we call multi-spec for short.

I first introduced a new mode `in-hint-decomposition`. The multi-spec behavior is only enabled when this mode is on, since the behavior is only acceptable for hint eval purposes.

Under this mode, when we are doing `UnionT ~> use_t` or `l ~> UnionT`, it will break up the union, calls `SpeculationKit.try_singleton_no_throws` introduced in D40170581 (https://github.com/facebook/flow/commit/7683900ad170f18c72f559db1efcd481876d0037). This will succeed normally and establish bound. When it fails, no bounds will be added and no errors would be added.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D40170579

fbshipit-source-id: a0f2b2d9ff1c84a0fc4aac6e1e2ac22e6b7d4c23