Allow forward references to enum definitions in types
Summary:
Currently enums are considered to just be normal variable declarations, and forward references to them are disallowed. However, similar to classes, they induce both types and values, and at the type level forward references to them should be allowed. This diff implements that for enums, exactly as works for classes.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40355006

fbshipit-source-id: cbaa3bd191221a11c2384e3779c58592344c2d9f