When reading ForTypeof from the env, don't allow forward references
Summary:
Currently when we refer to an undeclared variable with `ForTypeof`, we allow a forward reference and try to get the definition type for the variable. However, this behavior conflicts with that of the name_resolver and name_def_ordering, which assumes that a typeof type is a flow-sensitive read of the variable. This diff brings the env into alignment with the env builder by only allowing forward references to `ForType` lookups.

Changelog: [errors] The `typeof X` type annotation now respects the initialization state of `X`: if `X` hasn't been declared it will be an error to get its type.

Reviewed By: SamChou19815

Differential Revision: D40352824

fbshipit-source-id: f5b05a5c8045534abfaf80d3428f7f30b30504c7