Name def for this in functions where it's expected to be annotated
Summary:
This diff revises the approach to unannotated `this` in functions from D40194894 (https://github.com/facebook/flow/commit/5625ae0ebfda43fc9ca076f5b8b7289a7be5cea0). The issue there was that the `this` loc for a function without a this annotation still existed and could be read from, but wouldn't appear in the ordering and would never be resolved. To address this, while still not causing cycles to appear for functions with unannotated thises, we now treat such illegal `this`es as their own name def, which is resolved to `any` immediately in LTI mode and has no dependencies. For legacy mode, this `this` becomes an empty tvar, which will as previously be populated lazily.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40357221

fbshipit-source-id: cd2dcdc3ee9bdb8978920202564f1440efa307e8