[name-resolver] Fix logical assignment to member expressions
Summary:
Previously, logical op-assignment didn't quite work right when assigning to member expressions. The call to `add_projection` would immediately create an environment entry at the position of the LHS, and when we read from the LHS subsequently, we'll read that new projection rather than the original value of the LHS--so we're writing before we've read, and so the read of the LHS points to the write at the same location.

This diff removes the artificial projection and instead calls `assign_member` after performing the read and the refinement.

Reviewed By: SamChou19815

Differential Revision: D40391684

fbshipit-source-id: aae8ba9e2b0df25b12de7e1e81fa8a001112de9f