[reordered-env] Specialize synthesis of fully annotated function expression assigned to variables
Summary:
In cases like

```
const foo = (): number => foo();
```

It would be quite unfortunate to error on cycles, since we can easily synthesize a type for `foo` using the annotations on the function expression on the right hand.

In this diff, I recognized the pattern in name_def and name_def_ordering, and implemented the synthesis logic similar to fully annotated functions in env_resolution.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D40438411

fbshipit-source-id: ae149e73ff92c60e75540045b103e6a2f90acbb6