Do not any-propagate through FullyResolved tvars
Summary:
This diff does something simple but the implications are larger.

It's become clear that this approach for any-propagation will not be performant enough to power trusted types, so I'm repurposing this machinery to instead be used to constrain targs appropriately when `any` is flowed to some type structure with a deeply nested targ.

FullyResolved tvars cannot contain unresolved tvars, so there's no need to propagate through them.

Changelog: [Internal]

Reviewed By: SamChou19815

Differential Revision: D40476609

fbshipit-source-id: b4011a3cbccf41fc373732b34516657a037aa994