[cleanup] Do not use the reason of the default as init reason during variable assignments
Summary:
Currently, the use_op produced in destructuring.ml is only used in two places: destructuring assignments and destructuring in variable declaration. These use_ops are used for subtyping against providers, which needs an initializer to begin with, so it make much more sense to use the reason of initializers rather than the default.

With this change, we no longer need to record default in destructuring state.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D40489047

fbshipit-source-id: f0fce38e63505361183cc1c8ae761009033ec8b6