[implicit-instantiation] Roll back all errors except under-constrained-targ errors
Summary:
We should rollback all errors during implicit instantiation, because we will perform the same flow again with explicit targs later, so all errors we can get during the initial flow for the purpose of solving targs will also be emitted later. If we don't rollback these errors, we might double error.

However, the errors we emit when flowing targ solution to bound exist in the same function we emit `under-constrained-targ` errors. To overcome this, I wrap all final observer calls in a thunk, and only invoke it at the end of `solve_targs`.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D40403056

fbshipit-source-id: 6a1260e89fde99cc4c14d07500f2281663fbc8b4