Enable the `exact_empty_objects` option on by default
Summary:
A blog post will go out in concert with this change.

In previous work I already updated most of the tests to handle this change, so just a few changes here.

Changelog: [errors] Enable the `exact_empty_objects` option on by default. Look for a blog post for more details on how to handle this change.

Reviewed By: SamChou19815

Differential Revision: D40529053

fbshipit-source-id: 65e1798c0de1578ef3f9eb4b9df40d74044438d1