Rename mk_typeapp_instance to mk_typeapp_instance_annot
Summary:
Wrapping a TypeAppT in an annot is dangerous if the `ts` are non 0->1 tvars. I know of at least two places where that happens (addressed in subsequent diffs). I'm renaming this function in this diff to make it clear that the result is an AnnotT. Subsequent diffs introduce a version that does not wrap in an AnnotT and then replaces specific callsites. I'm splitting this change up into several diffs so that behavioral changes can be analyzed independently.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40527845

fbshipit-source-id: a7711ad02f2aaf93dee46db6ee576911dfd442c5