Use mk_typeapp_instance for TypeAppT subtyping
Summary:
These two callsites put potentially non 0->1 tvars behind an AnnotT. This makes the AnnotT not 0->1, which messes up invariants we rely on.

For my own purposes, this is blocking me from short-circuting on AnnotTs for any-propagation, which can cause timeouts due to extremely large mutually recursive types.

Instead of wrapping in an AnnotT, this diff performs the same behavior that an AnnotT would use. We slingshot for new lower bounds and reposition for new upper bounds. This maintains the current behavior of our test cases.

(we'll see what happens on the repos!)

Changelog: [Internal]

Reviewed By: SamChou19815

Differential Revision: D40527846

fbshipit-source-id: 51dcd300bea4484e6d2747b0dd0d0950e4dacaea