Do not any-propagate through AnnotTs
Summary:
We should not need to propagate through AnnotTs because AnnotTs are 0->1. When we find violations of this rule we should treat them as bugs and fix it.

Since AnnotTs are 0->1, there's no need to propagate any through them, as all the tvars will eventually resolve when the annotation resolves.

Reviewed By: SamChou19815

Differential Revision: D40527843

fbshipit-source-id: 0ce5b15d27eb265e09839c5ecc5e2e6151ccb356