Emit errors for each erroneous read from the environment
Summary:
Reordering shouldn't affect the behavior of programs in any way, other than just enabling them to be typechecked at all. However, due to the cacheing design of the new env, it's actually possible for information about the ordering to "leak" to the user, and a change in the ordering of the program can cause very different error behavior. This occurs for two classes of errors that are emitted during variable lookup in the new env: failed builtin lookups and use-before-declaration errors.

In both of these cases, the error is emitted at the point of searching for the variable in the environment. When we look something up in the environment, we cache the result, so that further reads of the same variable don't need to query the name resolver. However, this means that if a variable lookup results in an error, we'll only emit the error the first time we query the variable--and that's the first time *chronologically*, no the first time lexically. So for example:

```
function f() {
    // g();
    return x;
}

function g() {
    return x
}
```

Here, if `x` is an undefined variable, we'd expect to emit the error on the first use of `x` lexically, within `f()`. However, if we uncomment the call to `g` inside `f`, we'll now check `g()` first, and the error will move to `g()`. This will result in inconsistent behavior to users, uninituitive situations (a user could reasonably think that `x` must not be undefined in `f` since it's not an error there, only later) and errors that are difficult to suppress by the Flow team.

This diff addresses this issue by storing errors within the environment cache, and then repositioning and re-emitting the error every time the variable is looked up. This makes a number of functions, including both environment lookup and builtin lookup, produce a `result` of a type and optionally a list of errors to be emitted.

This does have a significant behavior change for undefined variable errors: previously, we would only emit an error for the first lookup of a variable, and then treat it as a global `any` typed variable. Now, we'll emit an error for *each* usage of the variable. (The same is true for use-before-define errors *now*, but previously to the rollout of the new env, they would be an error on each usage, and this diff returns us to that behavior: compare [this example](https://flow.org/try/#0B4bgUKYDYKYC4AJgILwICwCYRA) in Flow 0.190 and (e.g.) Flow 0.150.)

Changelog: [errors] `cannot-resolve-name` errors now are emitted for every use of an undefined variable, not just the first.

Reviewed By: samwgoldman

Differential Revision: D40401859

fbshipit-source-id: 6d132642c20462bf0623d0465c56db82de123c79