[implicit-instantiation] Ensure that we do not run more code in implicit instantiation than necessary
Summary:
In implicit instantiation, we do not flow targ to its upper bound. The reason is explained in D39482933 (https://github.com/facebook/flow/commit/6fc73672abbafcc9345282a4cbc184f5e77a1fa8). However, this flow is still important when we are not in implicit instantiation, and this diff ensures that we will not run unnecessary code in implicit instantiation, so that the flow won't be unexpectedly disabled.

The first obvious problem is that we are evaluating the return hint inside implicit instantiation mode, which is unnecessary and wrong. Therefore, I fixed that issue by moving the return hint eval code out.
However, with this change only, a regression test is still failing. It turns out that we are not resetting the flag `in_implicit_instantiation` properly. We might be doing `PolyT ~> CallT` in a speculation. In speculation, whenever we add_output, we will throw an exception, so the flag might never be reset! To fix this, I rewrite `Context.run_in_implicit_instantiation_mode` in a try-finally style so that the flag will always be reset properly

Reviewed By: jbrown215

Differential Revision: D40594530

fbshipit-source-id: 4351d9b8e290fd869c6618d6174f79602b91d622