Resolve empty arrays to Array<empty> under LTI and annotate more aggressively
Summary:
In order to allow empty arrays to be declared without automatically becoming errors, even if they're in positions that are never read from, this diff removes the EmptyArray error message on arrays that aren't on the RHS of variable declarations. It also now resolves the element type of these arrays to `empty`, rather than `any`: this enforces that these arrays are never written to and cannot be passed to another context that can write to them, while still allowing them to be passed into `$ReadOnlyArray<>` positions. Finally, this asks variable declarations to be annotated more aggressively, no longer requiring the --default-any parameter to be passed to annotate declarations with unknown element types like `var x = []`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40655769

fbshipit-source-id: 6890b95d5a84833ecba066ec37dbef863de0e1d0