Check defaults on annotations after resolving component
Summary:
Currently, an annotated parameter with a default is considered to be unresolvable. This is an issue for classes with methods with default arguments, because a parameter with a default depends on the tparams, which include the class itself. As a result we would have an unannotatable cycle in a case like `class C { m(x: number = 42) { } }`, because the default on `x` would depend on `C`.

Of course, in reality, a default on an annotation doesn't need to be resolved in the first place to resolve the type of the class. This diff addresses this by ensuring that annotations can be resolved without typechecking their defaults. We still need to check the default type against the annotation, but we now do this when checking function toplevels, at which point the annotation type should be resolved.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40689525

fbshipit-source-id: c7f9f6962fe109aafd2b262258d60a8065cec036