Add BuiltinName hint base
Summary:
I want to add a hint for jsx_fragment children to decompose the `children` prop from the `React$Fragment` type. In order to do this, I needed to make two changes:
1. (this diff) Add a BuiltinName base hint that uses builtins to turn it into a type
2. (next diff) Add the hint to jsx_fragment before visiting the children

This allows us to decompose hints from builtin types that are not referenced directly by name in the source code.

Reviewed By: SamChou19815

Differential Revision: D40642434

fbshipit-source-id: 1010b5af131b62047bfd148e7a2d96626a2d85e3