don't include body when printing module type
Summary:
in the places we currently use `Ty_printer` -- IDE stuff mostly -- we shouldn't print the body of a module type. similarly, we don't print the bodies of classes, interfaces or enums.

it turns out it's actually pretty good at printing module signatures, though. this could be useful for dumping `.js.flow` files. but that should opt into printing the bodies of all of the above structures.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D40722151

fbshipit-source-id: 4b1e7285bef8b20134bd9da1e45b3a36a05002b3