improve autocomplete kinds and details for declarations
Summary:
improves the autocomplete results for declarations (classes, interfaces, enums, modules, type aliases and variables).

the "label detail" (aka `lsp_detail`) appears in gray immediately after the name. it is used for function signatures and type aliases. the regular "details" appear off to the right or in an expanded details pane and are more verbose.

we were using the full "details" for "label details" in some declarations, so we ended up with `Foointerface Foo`. we don't need to include a label detail for this; the "kind" icon shows it's an interface and the full details still say "interface Foo".

this diff also improves the "kind" icon for type aliases: now instead of always being a "variable", it will reflect the type of the alias if it's a literal or an interface or a function.

Changelog: [internal]

Reviewed By: darichey

Differential Revision: D40708990

fbshipit-source-id: 895b4efdd3c8da575823f9dfbda78a57173f90d5