[contextual] Support overload resolution in a union in contextual typing
Summary:
When we are in contextual typing, trying to evaluate Instantiate_Call, we will inspect a type to see whether it is an intersection. If so, we will go into the overload resolution logic.

However, if the intersection is nested in a union, we will completely skip this step, and later down the decomposition, we will just choose the first branch in intersection, which is not what we want.

In this diff, I will inspect the callee type to see if it is a union, if so, I will break up the union, with each branch going to do its own overload resolution. This is important for cases like `(optionalArray || []).reduce(...)`, where we have union callee, and intersection polymorphic method type!

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D40704273

fbshipit-source-id: a45613652fd410d3dc9f9921fab88f6c4ef10e6b