include module type on import source in typed AST
Summary:
in `import ... from 'foo'`, this diff adds a type to `'foo'` in the typed AST. the type is the imported `ModuleT`, which is useful for a variety of things.

in particular, two immediate use cases are go-to-def on the string (we support this today but as a special case instead of just using the TAST) and autocompleting imports. for the latter, we need access to the `ModuleT` to know what to suggest as you do `import { | } from 'foo'`.

Reviewed By: panagosg7

Differential Revision:
D31957810 (https://github.com/facebook/flow/commit/6d64ccf11702e5d9646a3d41a7165f421855bb4a)

------------------------------------------------------------------------
(from b819098a3a2d0b986ba546d86e04a7277f6de4e9)

fbshipit-source-id: c1009bb75301d311ae9b04e60f08d949a26bd961