Move defaults back out of env resolution and remove from cycles
Summary:
This diff is, partially, a backing out of D39084949 (https://github.com/facebook/flow/commit/56ff97f080c5743c5f54970335e8a860fa463130) (cc SamChou19815), which moved evaluation of defaults from destructuring.ml and func_sig to env_resolution. The issue that has become apparent with that approach is that it means that the value of defaults must be fully resolved at the time that we're checking function or class signatures. In the case of defaults for parameters of class methods, this creates unavoidable and unbreakable cycles, because we assume that we must resolve all the tparams of the class--including `this`--in order to check the defaults. For example,

```
class C {
  m({foo = this}: {foo?: number}) { }
}
```
Here, we'll see `C` and `foo` as being in a cycle with each other, since in order to resolve `foo` we also must resolve its default, which depends on the class's `this`. But we don't actually need to do anything with the default here in order to resolve `C`'s signature--since this parameter is annotated, the default contributes nothing to either the signature nor to the type of `foo`. (This is not the case if the argument was contextual, and in that case we don't change the behavior here.)

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40738117

fbshipit-source-id: 32451363bbc7f83fc7160972527a8fa001290fdf