Allow predicate functions to be recursively resolved
Summary:
Previously, any predicate function could not be resolved recursively, whether or not it has an annotation, because we had to typecheck the predicate itself in order to determine the `OpenPredT`. However, we've already recorded in the name_resolver the predicates and writes that are involved in the predicate, so as long as the return type is annotated, we should be able to compute the OpenPredT without actually typechecking.

This diff implements this behavior. We now, for annotated predicate functions, can compute the predicate during env resolution without typechecking the predicate body.

One tricky bit here is that in order to support `instanceof` predicate functions, we also need to be able to typecheck `ExpressionDef` components recursively, since in a predicate like `function f(x: mixed): boolean %checks { return x instanceof C }`, the read of `C` is its own `ExpressionDef` name_def. This name_def is depended on by the predicates of the predicate function, so if the predicate is in a cycle, this name_def may be as well. In order to support this, we now allow simple, synthesizable expressions like identifiers to be recursively resolvable as well. We typecheck them slightly differently within env_resolution in order to not seal them too eagerly.

Reviewed By: SamChou19815

Differential Revision: D40730501

fbshipit-source-id: 5133cbcf30e23f63ead3d1fdc4084241e2f70bb6