Try to avoid emitting both recursive-definition and definition-cycle errors
Summary:
Due to the frequency of expression name_def nodes, it's common for definition-cycle and recursive-definition errors to coexist. For example, `var x = () => f(x)` would be both a definition-cycle pointing at `x`, and a cycle between the `Binding` name_def node for `x` and the ExpressionDef node for the argument expression. This diff attempts to resolve this by not emitting definition-cycle errors if all the nodes in the cycle are either `ExpressionDef`s or recursively defined, and if at least one node is recursively defined.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40769136

fbshipit-source-id: b6c2e54f9cb49ce7394cf059e586abac06555ff4