experimental: add scm saved state fetcher
Summary:
The "local" saved state fetcher looks for files called `.flow.saved_state` and `.flow.saved_state_file_changes` in the project dir. it's up to you to keep these files up to date, which makes this fetcher pretty much only useful for tests.

This diff adds a new "scm" saved state fetcher. You can generate multiple saved states (in the `.flow.saved_states` folder), and the fetcher will query git or hg to figure out which is the best one based on your current commit. it will also query for the files that have changed since that saved state, so you don't need a "file_changes" file.

No mechanism for populating the `.flow.saved_states` folder is provided. I'm still only using it for testing -- but this enables me to test different saved states while rebasing around. But I could imagine that directory as an NFS mount with a remote cron job populating saved states periodically for everyone to share.

NOTE: this is still experimental, and not documented in `--help` yet. In particular, the storage location may change (e.g. to /tmp/flow instead of being in the repo) and it doesn't yet obey the "mergebase-with" configs.

Changelog: [misc] Add an experimental "scm" saved state fetcher, which selects a saved state based on your currently source control revision.

Reviewed By: jbrown215

Differential Revision: D40651569

fbshipit-source-id: d12d7a31afc20f12e038e2d7d619dd0c9c5c1662