simplify error logic in solve_targs
Summary:
This diff simplifies the logic of `solve_targs` by removing the wrapping of the output of
`pin_types` in thunks. Instead, since we only care about implicit instantiation errors during
`pin_types`, we can reset errors before the call and filter out the errors of interest afterwards.
This way it's more clear exactly what errors we are interested in at that point.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40878250

fbshipit-source-id: 281c431ec6505e91890a7d3a9e0cf247a1d62afd