Soft fail on tarjan error
Summary:
Rather than producing a crash when Tarjan fails, we emit an internal error now. When this happens, we produce an empty list of components to be typechecked, which means that the module with the failure won't be typechecked at all. But this is still better than crashing!

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40901258

fbshipit-source-id: 5324c3310183523f756ed081ac1daec8b4bb4600