Soft fail in the environment
Summary:
This diff causes all tooling in environment construction to fail softly with an internal error rather than a crash. We raise an exception whenever issues come up, and in those cases skip typechecking the module entirely. Specifically, this diff replaces all calls to `failwith`, and it also wraps or replaces all calls to `Map.find` (for any kind of `Map`), since those raise `Not_found` errors when they miss.

This also addresses env.ml in a similar way (but there we raise errors as normal)

Reviewed By: SamChou19815

Differential Revision: D40906344

fbshipit-source-id: 8e58f05a048265092ce12611d0c3f79611eeb8fc