[lti] Declare dependency on synthesizable expressions for synthesis only nodes
Summary:
Currently, in the left to right algorithm, we don't declare dependency for argument appearing on the right of the argument we are trying to contextually type. In jsx instantiation, we don't declare dependency on all props and children. This is necessary for contextual typing, since we don't want to introduce cycles.

This approach reduces cycles, but sometimes it creates enforced environment errors. For example,

```
declare function poly<T>(T=>T, T): T;
() => {
   poly(v => v, foo1);
};
declare var foo1: string;
```

has an enforcement error because we are reading foo1 out of order. In general, we can't prevent that from happening for expressions that will use hints (fallback implemented in the next diff), but we can prevent this from happening for a subset of expressions that we know definitely won't require hint, like identifiers, members, etc. We can conservatively find such expressions syntactically, and then remove hints from such expressions. In name_def_ordering, we can then declare dependency on such expressions, so they can be checked before synthesis.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D40907186

fbshipit-source-id: 19a802e53b59040956b2972e957244a0550e72c8