Do not crash on invalid implicit instantiations
Summary:
It's possible to record a non-sensical implicit instantiation check, as in the added test case. Instead of crashing, we should just not run anything meaningful in the implicit instantiation.

This diff gets rid of the failwith for unexpected types and replaces it with a hardcoded default solution.

Note that there's already other errors emitted for calling a non-function, so we don't need to emit one here as well.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D40914264

fbshipit-source-id: c00958c41beea6c0cf7382d7ac3a094551dcebc4